+++ src/components/AgGridComponents/CustomAgGrid.tsx
@@ -9,6 +9,8 @@ import { CustomLoadingOverlay, CommonRenderer } from '../../components/AgGridCom
 import CustomFloatingFilter from '../../components/AgGridComponents/CustomAgGridFilter';
 import { orderBy } from 'lodash';
 import { checkStaticField, staticColumns } from "../../constants/columns"
+import { GridApi } from 'ag-grid-community';
+import { uniqBy } from 'lodash';
 
 export function reducer(state, action) {
   switch (action.type) {
@@ -147,12 +149,14 @@ export default function CustomAgGrid({
   actionLabel = null,
   actionEditable = false,
   onCellValueChanged = () => { },
-  showOnlyShowFilteredRecordSwitch = false
+  showOnlyShowFilteredRecordSwitch = false,
+  idProperty = "_id",
+  allowHeaderSelection = true
 }) {
   const [, setColumns] = useState(columns);
   const [columnApi, setColumnApi] = useState(null);
 
-  const [currentGridApi, setCurrentGridApi] = useState(null);
+  const [currentGridApi, setCurrentGridApi] = useState<GridApi | any>(null);
   const enableRowDrag = columns.some((d) => d.rowDrag);
 
   //  If you want to do something once grid binding done
@@ -170,7 +174,7 @@ export default function CustomAgGrid({
         if (oldSelectedRecords.length > 0) {
           params.api.forEachNode(function (node) {
             node.setSelected(
-              oldSelectedRecords.some((o) => o === node.data._id)
+              oldSelectedRecords.some((o) => o[idProperty] === node.data[idProperty])
             );
           });
         }
@@ -180,17 +184,22 @@ export default function CustomAgGrid({
     }
   };
 
-  useEffect(() => {
-    if (columnApi && loading === false) {
-      const columnState = JSON.parse(localStorage.getItem(renderedFrom));
-
-      if (columnState) {
+  const onFirstDataRendered = (e) => {
+    try {
+      if (localStorage.getItem(renderedFrom)) {
+        const columnState = JSON.parse(localStorage.getItem(renderedFrom));
         setTimeout(() => {
           columnApi.setColumnState(columnState);
-        }, 50)
+        }, 500)
       }
+    } catch (_) {
+      console.error("Error in configuring columns on onFirstDataRendered method")
     }
-  }, [columnApi, loading])
+
+    if (currentGridApi) {
+      currentGridApi.sizeColumnsToFit()
+    }
+  }
 
   const onColumnMoved = (params) => {
     if (params?.source === "uiColumnDragged") {
@@ -210,7 +219,7 @@ export default function CustomAgGrid({
     if (currentGridApi && selectedRecords.length) {
       currentGridApi.forEachNode(function (node) {
         node.setSelected(
-          selectedRecords.some((o) => o._id === node.data._id)
+          selectedRecords.some((o) => o[idProperty] === node.data[idProperty])
         );
       });
     }
@@ -221,7 +230,7 @@ export default function CustomAgGrid({
         if (oldSelectedRecords.length > 0) {
           currentGridApi.forEachNode(function (node) {
             node.setSelected(
-              oldSelectedRecords.some((o) => o === node.data._id)
+              oldSelectedRecords.some((o) => o[idProperty] === node.data[idProperty])
             );
           });
         }
@@ -240,19 +249,6 @@ export default function CustomAgGrid({
     // trimInput: true,
     // debounceMs: 1000,
   };
-
-  const getWidth = (field, columnWidth) => {
-    if (localStorage.getItem(renderedFrom)) {
-      const storedColumns = JSON.parse(localStorage.getItem(renderedFrom));
-
-      const indexOfField = storedColumns.findIndex((d) => d.colId === field);
-      if (indexOfField > -1) {
-        return storedColumns[indexOfField].width;
-      }
-      return columnWidth;
-    }
-    return columnWidth;
-  }
 
   const getActionColumn = () => {
     if (allowAction) {
@@ -285,8 +281,8 @@ export default function CustomAgGrid({
         cellRenderer={column.cellRenderer ?? null}
         cellRendererParams={column.cellRendererParams ?? null}
         minWidth={column.width ?? 180}
-        width={getWidth(column.field, column.width) ?? 180}
-        flex={1}
+        // width={getWidth(column.field, column.width) ?? 180}
+        // flex={1}
         rowDrag={column.rowDrag ?? false}
         hide={staticColumns.indexOf(column.field) >= 0 ? checkStaticField(renderedFrom, column.field) :
           (column.hasOwnProperty("show") && !column?.show) ? true : false}
@@ -307,8 +303,8 @@ export default function CustomAgGrid({
         cellRenderer={column.cellRenderer ?? null}
         cellRendererParams={column.cellRendererParams ?? null}
         minWidth={column.width ?? 180}
-        width={getWidth(column.field, column.width) ?? 180}
-        flex={1}
+        // width={getWidth(column.field, column.width) ?? 180}
+        // flex={1}
         filterParams={customFilterParams}
         hide={(column.hasOwnProperty("show") && !column?.show) ? true : false}
         comparator={() => {
@@ -347,6 +343,7 @@ export default function CustomAgGrid({
 
           <div className="ag-theme-material ag-grid-listing-grid" style={{ zIndex: -500, position: 'inherit' }}>
             <AgGridReact
+              onFirstDataRendered={onFirstDataRendered}
               gridOptions={customGridOptions}
               rowData={dataRows}
               onColumnMoved={onColumnMoved}
@@ -378,6 +375,8 @@ export default function CustomAgGrid({
                 floatingFilter: true,
                 sortable: true,
                 suppressMenu: true,
+                // suppressSizeToFit: true,
+                // suppressAutoSize: true,
                 // headerCheckboxSelection: true,
                 // checkboxSelection: true,
                 floatingFilterComponentParams: { suppressFilterButton: true }
@@ -428,14 +427,14 @@ export default function CustomAgGrid({
                   try {
                     let oldSelectedRecords = localStorage.getItem(`${renderedFrom}_selected`) ? JSON.parse(localStorage.getItem(`${renderedFrom}_selected`)) : []
 
-                    if (event.node.isSelected() === true && !oldSelectedRecords.some(s => s === event.node.data.id)) {
-                      oldSelectedRecords = [...oldSelectedRecords, event.node.data.id];
+                    if (event.node.isSelected() === true && !oldSelectedRecords.some(s => s[idProperty] === event.node.data[idProperty])) {
+                      oldSelectedRecords = [...oldSelectedRecords, event.node.data];
                       localStorage.setItem(`${renderedFrom}_selected`, JSON.stringify(oldSelectedRecords));
                     }
                     else if (event.node.isSelected() === false) {
 
                       if (oldSelectedRecords.length > 0) {
-                        localStorage.setItem(`${renderedFrom}_selected`, JSON.stringify(oldSelectedRecords.filter(f => f !== (event.data._id ?? event.data.id))));
+                        localStorage.setItem(`${renderedFrom}_selected`, JSON.stringify(oldSelectedRecords.filter(f => f[idProperty] !== event.data[idProperty])));
                       }
                     }
                   } catch (ex) {
@@ -454,10 +453,10 @@ export default function CustomAgGrid({
                   try {
                     let oldSelectedRecords = localStorage.getItem(`${renderedFrom}_selected`) ? JSON.parse(localStorage.getItem(`${renderedFrom}_selected`)) : []
                     if (oldSelectedRecords.length > 0) {
-                      const uniqueRecords = [...new Set([...oldSelectedRecords, ...event.api.getSelectedRows().map(m => m._id)])]
+                      const uniqueRecords = uniqBy([...oldSelectedRecords, ...event.api.getSelectedRows()], idProperty)
                       localStorage.setItem(`${renderedFrom}_selected`, JSON.stringify(uniqueRecords));
                     } else {
-                      localStorage.setItem(`${renderedFrom}_selected`, JSON.stringify(event.api.getSelectedRows().map(m => m._id)))
+                      localStorage.setItem(`${renderedFrom}_selected`, JSON.stringify(event.api.getSelectedRows()))
                     }
                   } catch (ex) {
                     console.error("Error in getting / storing selected records")
@@ -471,7 +470,7 @@ export default function CustomAgGrid({
               }}
               immutableData={true}
               getRowNodeId={(data) => {
-                return data._id ?? data.id;
+                return data[idProperty];
               }}
               pagination={allowPagination}
               suppressPaginationPanel={true}
@@ -484,7 +483,7 @@ export default function CustomAgGrid({
                   filter={false}
                   pinned="left"
                   lockPinned={true}
-                  headerCheckboxSelection={true}
+                  headerCheckboxSelection={allowHeaderSelection}
                   headerCheckboxSelectionFilteredOnly={true}
                   checkboxSelection={true}
                   resizable={false}