+++ src/pages/ProductInventory/ProductInventoryDetailsPage.tsx
@@ -404,7 +404,7 @@ const ProductInventoryDetailsPage = () => {
                 alignItems="center"
               >
                 <Typography variant="subtitle2">
-                  BOM
+                  BOM - Product
                 </Typography>
               </Box>
               {(