+++ src/pages/RentalManagement/ManageAdditionalCostDialog.tsx
@@ -10,6 +10,15 @@ import CustomDialogFooter from '../../components/CustomDialog/CustomDialogFooter
 import CustomButton from '../../components/Helpers/CustomButton'
 import { CustomDialogTransition } from '../../constants/helpers'
 import ConfirmCancelDialog from '../../components/ConfirmCancelDialog'
+import { object, string, number } from "yup";
+
+const AdditionalCostSchema = object().shape({
+    type: string().required("Please select cost type"),
+    description: string().required("Please add description"),
+    qty: number().min(1, "Please add quantity").required("Please add quantity"),
+    uom: string().required("Please add UOM").nullable(),
+    amount: number().min(1, "Please add amount").required("Please add amount"),
+});
 
 export default function ManageAdditionalCostDialog({ open, currencySymbol, isNew, record, costTypeList, uomTypeList, onClose, onSubmit }) {
 
@@ -42,6 +51,7 @@ export default function ManageAdditionalCostDialog({ open, currencySymbol, isNew
                 />
 
                 <Formik
+                    validationSchema={AdditionalCostSchema}
                     initialValues={record}
                     enableReinitialize={true}
                     onSubmit={onSubmit}
@@ -74,52 +84,60 @@ export default function ManageAdditionalCostDialog({ open, currencySymbol, isNew
                                                             newValue
                                                         );
                                                     }}
-
                                                     renderInput={(params) => <TextField
                                                         {...params}
                                                         variant="outlined"
                                                         name="nameField"
                                                         label="Cost Type"
+                                                        required={true}
+                                                        margin="dense"
+                                                        error={touched["type"] && Boolean(errors["type"])}
+                                                        helperText={touched["type"] && errors["type"]}
                                                     />}
                                                 />
                                             </Grid>
 
                                             <Grid item xs={12} sm={12} md={6}>
-                                                <FormControl fullWidth variant="outlined" size="small">
-                                                    <InputLabel htmlFor="description">Description</InputLabel>
-                                                    <OutlinedInput
-                                                        id="description"
-                                                        type="text"
-                                                        name="description"
-                                                        value={values.description}
-                                                        onChange={(e) => {
-                                                            setFieldValue(
-                                                                "description",
-                                                                e.target.value
-                                                            );
-                                                        }}
-                                                    />
-                                                </FormControl>
+                                                <TextField
+                                                    variant="outlined"
+                                                    type="text"
+                                                    label="Description"
+                                                    required={true}
+                                                    name="description"
+                                                    fullWidth
+                                                    margin="dense"
+                                                    value={values.description}
+                                                    error={touched["description"] && Boolean(errors["description"])}
+                                                    helperText={touched["description"] && errors["description"]}
+                                                    onChange={(e) => {
+                                                        setFieldValue(
+                                                            "description",
+                                                            e.target.value
+                                                        );
+                                                    }}
+                                                />
                                             </Grid>
 
 
                                             <Grid item xs={12} sm={12} md={6}>
-                                                <FormControl fullWidth variant="outlined" size="small">
-                                                    <InputLabel htmlFor="quantity">Quantity</InputLabel>
-                                                    <OutlinedInput
-                                                        id="quantity"
-                                                        type="text"
-                                                        // size="small"
-                                                        name="quantity"
-                                                        value={values.qty}
-                                                        onChange={(e) => {
-                                                            setFieldValue(
-                                                                "qty",
-                                                                e.target.value.replace(/[^0-9]/g, '')
-                                                            );
-                                                        }}
-                                                    />
-                                                </FormControl>
+                                                <TextField
+                                                    variant="outlined"
+                                                    type="number"
+                                                    label="Quantity"
+                                                    required={true}
+                                                    name="quantity"
+                                                    fullWidth
+                                                    margin="dense"
+                                                    value={values.qty}
+                                                    error={touched["qty"] && Boolean(errors["qty"])}
+                                                    helperText={touched["qty"] && errors["qty"]}
+                                                    onChange={(e) => {
+                                                        setFieldValue(
+                                                            "qty",
+                                                            e.target.value.replace(/[^0-9]/g, '')
+                                                        );
+                                                    }}
+                                                />
                                             </Grid>
 
                                             <Grid item xs={12} sm={12} md={6}>
@@ -137,34 +155,43 @@ export default function ManageAdditionalCostDialog({ open, currencySymbol, isNew
                                                             newValue
                                                         );
                                                     }}
-
                                                     renderInput={(params) => <TextField
                                                         {...params}
                                                         variant="outlined"
                                                         name="nameField"
                                                         label="UOM"
+                                                        margin="dense"
+                                                        error={touched["uom"] && Boolean(errors["uom"])}
+                                                        helperText={touched["uom"] && errors["uom"]}
                                                     />}
                                                 />
                                             </Grid>
 
 
                                             <Grid item xs={12} sm={12} md={6}>
-                                                <FormControl fullWidth variant="outlined" size="small">
-                                                    <InputLabel htmlFor="amount">Amount</InputLabel>
-                                                    <OutlinedInput
-                                                        id="amount"
-                                                        startAdornment={currencySymbol ? <InputAdornment position="start">{currencySymbol}</InputAdornment> : ""}
-                                                        type="text"
-                                                        name="amount"
-                                                        value={values.amount}
-                                                        onChange={(e) => {
-                                                            setFieldValue(
-                                                                "amount",
-                                                                e.target.value.replace(/[^0-9]/g, '')
-                                                            );
-                                                        }}
-                                                    />
-                                                </FormControl>
+                                                <TextField
+                                                    variant="outlined"
+                                                    id="amount"
+                                                    type="number"
+                                                    label="Amount"
+                                                    required={true}
+                                                    name="amount"
+                                                    fullWidth
+                                                    margin="dense"
+                                                    value={values.amount}
+                                                    error={touched["amount"] && Boolean(errors["amount"])}
+                                                    helperText={touched["amount"] && errors["amount"]}
+                                                    onChange={(e) => {
+                                                        setFieldValue(
+                                                            "amount",
+                                                            e.target.value.replace(/[^0-9]/g, '')
+                                                        );
+                                                    }}
+                                                    InputProps={{
+                                                        startAdornment: <InputAdornment position="start">{currencySymbol}</InputAdornment>
+                                                    }}
+                                                />
+
                                             </Grid>
 
                                         </Grid>